day1/requirements.txt
# Core LangChain and OpenAI
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.5
langchain-milvus==0.1.5
langchain-text-splitters==0.3.2
openai==1.54.3

# Vector Stores
faiss-cpu==1.9.0.post1
pymilvus==2.4.8

# PDF Processing
PyPDF2==3.0.1
pypdf==5.1.0
reportlab==4.2.5

# Web Interface
streamlit==1.40.1

# FastAPI Backend
fastapi==0.115.5
uvicorn==0.34.0
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
